+++ src/Forms/Components/Fieldset.php
@@ -21,7 +21,7 @@ class Fieldset extends Component
      */
     protected array $schema = [];
 
-    protected int $columns = 12;
+    protected int $columns = 1;
 
     public function __construct(mixed $label = null)
     {