+++ src/Forms/Components/Step.php
@@ -27,7 +27,7 @@ class Step extends Component
      */
     protected array $schema = [];
 
-    protected int $columns = 12;
+    protected int $columns = 1;
 
     public function __construct(mixed $label)
     {