+++ src/Forms/Components/Tab.php
@@ -22,7 +22,7 @@ class Tab extends Component
      */
     protected array $schema = [];
 
-    protected int $columns = 12;
+    protected int $columns = 1;
 
     public function __construct(mixed $label)
     {